checkers/requirements.txt
requests==2.27.1
pwntools==4.8.0
ed25519-blake2b==1.4
python-socketio==5.6.0
websocket-client==1.3.2
gmpy2==2.1.2
pycryptodome==3.14.1
beautifulsoup4==4.6.0
